+++ .eslintrc.js
@@ -379,6 +379,7 @@ module.exports = {
           'resolver',
           'resolvers',
           'respawn',
+          'Retryable',
           'rgba',
           'rmdir',
           'rollbacking',