+++ packages/modules/module-replication/src/api-handlers/pause_replication.js
@@ -1,0 +1,7 @@
+const handler = async (req, res) => {
+  await req.resolve.eventstoreAdapter.setReplicationPaused(true)
+  res.status(200)
+  res.end()
+}
+
+export default handler