+++ packages/runtime/adapters/eventstore-adapters/eventstore-lite/src/load-secrets.ts
@@ -9,15 +9,25 @@ const loadSecrets = async (
   { database, escapeId, escape, secretsTableName }: AdapterPool,
   filter: SecretFilter
 ): Promise<SecretsWithIdx> => {
-  const { idx, limit, skip } = filter
+  const { idx, limit, skip, ids, includeDeleted } = filter
 
   const tableNameAsId: string = escapeId(secretsTableName)
   const searchIdx: number = idx == null ? 0 : idx
   const skipRows = skip === undefined ? 0 : skip
 
+  if (ids && ids.length === 0) {
+    return {
+      secrets: [],
+      idx: searchIdx,
+    }
+  }
+
   const rows = await database.all(
     `SELECT idx, id, secret FROM ${tableNameAsId}
-    WHERE secret IS NOT NULL AND idx >= ${+searchIdx}
+    WHERE ${
+      !includeDeleted ? `"secret" IS NOT NULL AND` : ''
+    } idx >= ${+searchIdx}
+    ${ids ? `AND id IN (${ids.map((id) => escape(id)).join(',')})` : ''}
     ORDER BY "idx" ASC
     LIMIT ${+limit} OFFSET ${skipRows}`
   )