+++ packages/runtime/adapters/readmodel-adapters/readmodel-mysql/src/index.ts
@@ -12,6 +12,7 @@ import type {
   AdapterPool,
   AdapterOptions,
   AdapterApi,
+  AdminPool,
 } from './types'
 
 import buildUpsertDocument from './build-upsert-document'
@@ -42,6 +43,9 @@ import pause from './pause'
 import resume from './resume'
 import status from './status'
 import build from './build'
+
+import _createResource from './resource/create'
+import _destroyResource from './resource/destroy'
 
 const store: CurrentStoreApi = {
   defineTable,
@@ -100,3 +104,20 @@ const createAdapter = _createAdapter.bind<
 >(null, implementation)
 
 export default createAdapter
+
+const pool = {
+  connect,
+  disconnect,
+  escapeId,
+  escapeStr,
+} as AdminPool
+
+const createResource = _createResource.bind(null, pool)
+const destroyResource = _destroyResource.bind(null, pool)
+
+Object.assign(pool, {
+  createResource,
+  destroyResource,
+})
+
+export { createResource as create, destroyResource as destroy }